// IFG-BaseProject/extensions/ifg-tools/src/models/requirements/RequirementType.ts
export enum RequirementType {
    Date = 'date',
    Level = 'level',
    LevelId = 'levelId',
    Inventory = 'inventory',
    PropInNode = 'propInNode',
    PropHasTagInNode = 'propHasTagInNode',
    Dialogue = 'dialogue',
    ChapterComplete = 'chapterComplete',
    TaskComplete = 'taskComplete',
    AccountRequirement = 'account',
    Tutorial = 'tutorial',
    QuickPlaySessions = 'quickPlaySessions'
}